created components folder and created export import logic for react components
